Split SubstituteExtensions into multiple files
